Begin the Lecture 6 notes on version control
Event though I probably don't need to make a commit message this
long, I'm putting into practice the techniques taught in the course,
which dictate that good commit messages start with a 50 character
summary followed by a blank line and then a more descriptive message
if necessary. This one, maybe not so much, but pretty cool to do.